Fix travis building issue

(cherry picked from commit 1cbcdf55b0fbc09df92c055cae8ee6d2f2131879)

+++ src/rpcmisc.cpp
@@ -144,9 +144,9 @@ UniValue getstateinfo(const UniValue& params, bool fHelp)
     LuxDGP luxDGP(globalState.get());
 
     UniValue obj(UniValue::VOBJ);
-    obj.push_back(Pair("maxblocksize", (int64_t)luxDGP.getBlockSize(chainActive.Height())));
-    obj.push_back(Pair("blockgaslimit", (int64_t)luxDGP.getMinGasPrice(chainActive.Height())));
-    obj.push_back(Pair("mingasprice", (int64_t)luxDGP.getBlockGasLimit(chainActive.Height())));
+    obj.push_back(Pair("maxblocksize", (int)luxDGP.getBlockSize(chainActive.Height())));
+    obj.push_back(Pair("blockgaslimit", (int)luxDGP.getMinGasPrice(chainActive.Height())));
+    obj.push_back(Pair("mingasprice", (int)luxDGP.getBlockGasLimit(chainActive.Height())));
 
     return obj;
 }
